requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asttokens
async-lru
attrs
babel
backoff
bcrypt
beautifulsoup4
bleach
build
cachetools
certifi
cffi
charset-normalizer
chromadb
click
colorama
coloredlogs
comm
dataclasses-json
debugpy
decorator
defusedxml
diskcache
distro
Django
djangorestframework
durationpy
executing
fastjsonschema
filelock
flatbuffers
fqdn
frozenlist
fsspec
google-auth
googleapis-common-protos
gpt4all
greenlet
grpcio
h11
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
ipykernel
ipython
ipython_pygments_lexers
isoduration
jedi
Jinja2
jiter
joblib
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-ollama
langchain-text-splitters
langchain-together
langgraph
langgraph-checkpoint
langgraph-checkpoint-sqlite
langgraph-prebuilt
langgraph-sdk
langsmith
llama_cpp_python
markdown-it-py
MarkupSafe
marshmallow
matplotlib-inline
mdurl
mistune
mmh3
mpmath
multidict
mypy_extensions
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numpy
oauthlib
ollama
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
ormsgpack
overrides
packaging
pandocfilters
parso
platformdirs
posthog
prometheus_client
prompt_toolkit
propcache
protobuf
psutil
pure_eval
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-dotenv
python-json-logger
pywin32
pywinpty
PyYAML
pyzmq
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rpds-py
rsa
scikit-learn
scipy
Send2Trash
shellingham
six
sniffio
soupsieve
SQLAlchemy
sqlite-vec
sqlparse
stack-data
sympy
tenacity
terminado
threadpoolctl
tinycss2
tokenizers
tornado
tqdm
traitlets
typer
types-python-dateutil
typing-inspect
typing-inspection
typing_extensions
tzdata
uri-template
urllib3
uvicorn
watchfiles
wcwidth
webcolors
webencodings
websocket-client
websockets
xxhash
yarl
zipp
zstandard